Handle undefined config when not specified
`react-axe` will fail with an exception if the configuration object is
not specified in the call to `axe()`:

A call like this:

```es6
axe(React, ReactDOM, 3000)
```

Throws an error like this:

```
index.js:374 Uncaught TypeError: Cannot read property 'runOnly' of undefined
    at reactAxe (index.js:374)
```

The existence of the configuration object is checked lower down in the
function, but the early check to see if `runOnly` is relevant doesn't work if
the object isn't there.

A workaround is to specify an empty object `{}`, but it seems from the
type definitions that this is truly meant to be an optional parameter.

Fix #176